Add *.pyc to .gitignore
In theory, this is not needed, since in Python 3, the pyc files are in __pycache__.

However, at the SF Django Girls event, several participants used Python 2, since
they had that already installed and it worked for the tutorial--up to this point, where
it broke in a hard-to-diagnose way. Their local_settings.pyc file was getting to
Heroku and it was trying to use local settings in deployment.

Adding *.pyc to .gitignore is harmless for people using Python 3, and solves this
problem for people using Python 2.